steam window and harvesters update
This implements showing of the buffer stream (makes it a little slower) and an update to the harvesters. Unfortunately the harvesters are still just the same as the cv2 pipeline, just more verbose. The thing that we are looking for is to be able to do a "binning" on the pixel ON THE DEVICE and speed up the acquision frame rate

+++ camera/camera_control.py
@@ -1,7 +1,9 @@
 import cv2, os, tqdm, glob, time, datetime
 import matplotlib.pyplot as plt
 
-def imshow_resize(frame_name = "img", frame = 0, resize_size = [640,480], default_ratio = 1.3333, always_on_top = True, use_waitkey = True):
+# cv2 imshow but resized to a size that can fit in a normal monitor size
+def imshow_resize(frame_name = "img", frame = 0, resize_size = [640,480], default_ratio = 1.3333, 
+                  always_on_top = True, use_waitkey = True, move_to = [1920,1]):
 
     frame = cv2.resize(frame, dsize=resize_size)
     cv2.imshow(frame_name, frame)
@@ -9,37 +11,50 @@ def imshow_resize(frame_name = "img", frame = 0, resize_size = [640,480], defaul
         cv2.setWindowProperty(frame_name, cv2.WND_PROP_TOPMOST, 1)
     if use_waitkey:
         cv2.waitKey(1)
-    cv2.moveWindow(frame_name,1920-resize_size[0],1)
+    cv2.moveWindow(frame_name,move_to[0]-resize_size[0],move_to[1])
     return True
 
-def del_dir_contents(path_to_dir):
-    files = glob.glob(os.path.join(path_to_dir,'*'))
-    for f in files:
-        os.remove(f)
+# this deletes all the dir contents, recursive or not 
+def del_dir_contents(path_to_dir, recursive = False): 
+    if recursive:
+        files = glob.glob(os.path.join(path_to_dir,'**/*'), recursive=recursive)
+        for f in files:
+            if not os.path.isdir(f):
+                os.remove(f)
+    else: # default
+        files = glob.glob(os.path.join(path_to_dir,'*'))
+        for f in files:
+            os.remove(f)
 
-def capture_images_for_time(cap,N):
-
+# this is a buffer that continues to capture images until the the specified time has elapsed 
+# this is because if you use time.sleep() the image taken is buffered and not at the actual elapsed time
+def capture_images_for_time(cap,N, show_images = False, move_to = [100,100], resize_size = [640,480]):
     start_time = datetime.datetime.now()
-
     while True:
         current_time = datetime.datetime.now()
         ret, frame = cap.read()
         if not ret:
             print("Error: Unable to capture frame.")
             break
-        # imshow_resize("img", frame, resize_size=[640,480])
+        if show_images:
+            imshow_resize("stream", frame, resize_size=resize_size, move_to=move_to)
         if start_time + datetime.timedelta(0,N + 0.1) < current_time:
             break 
 
+# this captures the first N images to clear the pipeline (sometime just black images)
 def clear_camera_image_buffer(cap,N=2):
     for i in range(N):
         ret, frame = cap.read()
 
-def simple_capture_data(camera_settings, plate_parameters = None, testing = False):
+def simple_capture_data(camera_settings, plate_parameters = None, testing = False, output_dir = None):
 
     todays_date = datetime.date.today().strftime("%Y-%m-%d")
 
-    output_dir = os.path.join(os.getcwd(),'output',plate_parameters['experiment_name'],plate_parameters['plate_name'],todays_date)
+    if output_dir == None:
+        output_dir = os.path.join(os.getcwd(),'output',plate_parameters['experiment_name'],plate_parameters['plate_name'],todays_date)
+    else:
+        output_dir = os.path.join(output_dir,plate_parameters['experiment_name'],plate_parameters['plate_name'],todays_date)
+    
     os.makedirs(output_dir,exist_ok=True)
     if testing:
         del_dir_contents(output_dir)
@@ -68,7 +83,10 @@ def simple_capture_data(camera_settings, plate_parameters = None, testing = Fals
     text_x = (cam_width - text_size[0]) // 2  # Center horizontally
     text_y = 250  # 250 pixels from the top
 
-    time_between_images_seconds = 0.1
+    # time_between_images_seconds = 0 # this is just for testing 
+    img_file_format = 'png' # slow and lossless but smaller 
+    # # img_file_format = 'jpg' # fast but lossy small files
+    # # img_file_format = 'bmp' # fastest and lossess huge files
 
     # Open the camera0
     cap = cv2.VideoCapture(int(camera_id))
@@ -95,7 +113,7 @@ def simple_capture_data(camera_settings, plate_parameters = None, testing = Fals
         image_name = current_time + '.' + img_file_format
         image_filename = os.path.join(output_dir, image_name)
 
-        # cv2.imwrite(image_filename, frame)
+        cv2.imwrite(image_filename, frame)
         # print(f"\nCaptured image {i+1}/{num_images}")
 
         # Put the text on the image
@@ -110,7 +128,7 @@ def simple_capture_data(camera_settings, plate_parameters = None, testing = Fals
             if delay_time<0:
                 pass
             else:
-                capture_images_for_time(cap,delay_time)
+                capture_images_for_time(cap,delay_time, show_images=True,move_to = [1920,520])
         # time.sleep(1)
 
     # Release the camera
